Move redirect info to its own endpoint
- Create new /api/redirects GET endpoint that returns redirect slugs
- Remove redirect data from /api/ping endpoint response
- Add RedirectsResponseSchema and remove redirects from PingResponseSchema
- Implement same caching logic in new endpoint
- Remove caching functionality from ping endpoint

Implements DIO-140
